Return null on 404s for guides and users
When testing this on guides that didn't exist, the graphql playground
was left hanging on a response when the backend had encountered an
exception from axios.

This uses a promiseAllOrError work around for Promise.all that doesn't
short circuit if one of the request fails.

This also doesn't behave differently between a 404, a 403, or a 500,
so this could be improved in the future.